modified:   include/mm/paging.h The main (and for now the only one) page global directory is made accessible to modules that include the paging header.
	modified:   include/mm/pgdir.h
The page directory struct is made page-aligned.

	modified:   include/mm/pgtable.h
The page table struct is made page-aligned.

	modified:   kernel/kernel.c
Set up and initialized paging.

	modified:   kernel/mm/paging.c
Created one page table, and used it to map the first 4 MB of the
physical address space in paging_startup(), in addition to setting cr3
with the address of the main PGD.
Also changed the order of registers in assembly in switch_pg_dir() to
follow AT&T syntax.